requirements.txt
Flask==2.3.2
hnswlib==0.7.0
nanoid==2.0.0
numpy==1.24.3
openai==0.27.7
python-dotenv==1.0.0
Requests==2.31.0
streamlit==1.23.1
streamlit_chat==0.0.2.2
torch==2.0.1
transformers==4.29.2
